07 13 2025
It is going to take a while but could I automate this?

what kind of script

could I use a window function?

I don't know.

maybe I can write a query to break in the first space.

so, it reads the fist spaces of a line until it reaches the first open space.

it breaks that into NUMBER.