07 22 2025
just continuing to work on the COUNTRY table.

it is very long.  yeah, maybe like in 5 more days?

it is taking 40 lines or so per pomodoro.

I will have to see if I can make this faster somehow?

I doubt it for this one.

I will just have to keep at it.

yea