07 31 2025
ok!

it more or less took me 5 days to place commas appropriately on the COUNTRY.txt file

I am glad to be done and I am excited to move on to checking my work the next time I work on this.

I am getting a better sense of what joins are already