08 25 2025
Having weird problems.

there is a type mismatch.

I am just following the description given of that table contents in the problem description page.

I may have to change those around to fit the suggestions in the Review Data Type Mapping" dialog box.

I so agree with the suggestions since why is INDEPYEAT a VARCHAR2?  why isn't it just a NUMBER?  DUH.