08 26 2025
Another day another frustration

I am now just again not sure why the column INDEPYEAR isn't working

but I have a new lead I guess

I thought it might have been because the INDEPYEAR for JPN is -066?  I checked with the original table and it adds up.  it says -66 or whatever.  so, i have to go with that.