08 31 2025
I need to practice a join.

so, I may try to solve this on the github screen and not download it here.

I need to get a hang of joins really quick for the possible incoming interview.

But I may have time either way.

The join order I need to review.

But I have done JOINs in the past.
Just need to refresh.

I think breaking this down will be great.